server: implement the ban command
to implement this command, data structures had to changed. namely the simple
SOCKET_LIST had to be changed to a dict where the key is the addr of the socket
and the value is the socket itself. also a OP_LIST has been added which contains
all the addr of clients that can perform op actions such as ban itself.
also the old send function is now called send_all.